Only redeploy once.
Redeploying the original bundle fails when Jdisc is modified to
allow duplicate bundles. The original bundle will be invisible
until it has been uninstalled along with destructing the
component for the initial application. Re-installing it before
that point will not help, because installing an already installed
bundle is a NOP.

Alternatively, we could have added a delay between redeploys,
larger than the deconstruct delay of 60 sec, but this seems
unnecessary.