Use realpath instead of readlink
Say /foo/slink is a symlink with content "target", and /foo/target is a file.
Then the symlink is valid and points to /foo/target.

However, if PWD is /bar and the symlink content is read, that content "target"
is NOT a valid file as there is no /bar/target file.  This is what the current
code does.  If the content of the symlink had been an absolute path name
(/foo/target), it WOULD be valid, which is why it works today.