Dump status page to help understand why resume completes immediately
Also, test_upgrade_of_downed_node waits until the state is 'down', instead of
one of "sdm" and waiting 30s.

Also, when waiting for the cluster controller to ack a storage node being
stopped, we used to wait until the the node were in one of the states
'stopping', 'maintenance', or 'down'.  However a storage node in 'stopping' is
a transitional state that starts with the node behaving as-if 'up', which the
client probably does not want.  Therefore 'stopping' is removed from this set.