Add an implicit barrier for buckets pending background indexing
This helps avoid a race condition in tests where the test framework
erroneously believes all maintenance ops are done since the distributor
does not _directly_ indicate any buckets have pending maintenance ops.

Count the number of bucket replicas that _should_ (eventually) be indexed
and made searchable according to the ideal state, but where this currently
is not the case. This works as an implicit barrier to wait for search nodes
to asynchronously complete background bucket moving and search node to
distributor bucket info change notifications.

This only works reliably as a barrier for non-grouped clusters, as this
code does not have enough knowledge of the group topology and ideal state
to make an informed decision on what _other_ replicas than the first one
should be ready.