Handle case where CC heartbeat response is sent between feed ops in test
Test feeds 2 docs to independent buckets, and expects that the bucket
count reported back to the CC is initially 0, then 2. This does not hold
true if the bucket DB state is sampled and sent back to the cluster
controller _after_ the first document has been fed but _before_ the
second document has been fed, in which case 1 bucket will be reported.

Test has been relaxed to accept intermediate values before settling
on 2 buckets reported.