Enable FileStor debug logging for node retirement test
On rare occasions this test fails with empty, orphaned buckets on
one of the content nodes (usually when run under Valgrind and things
run with the speed and grace of a gazelle stuck in mud). Hypothesis
is that some (possibly resent) operation is triggering an implicit
bucket creation in the DB; debug logging will hopefully point out
the source.