Selectively enable debug logging around initial feed
Test fails with (presumed transient) reduced coverage once in a
blue moon. This is most likely an instance of nodes "crossing the
streams" due to async bucket info changes being sent from the
content nodes due to readiness-changes at the same time as the
distributor itself mutates the bucket metadata, but want to catch
it in the act to make sure.

Only observed for the initial feed round, so likely related to
replica bootstrapping.