Migrate workflow "systemtest-base-el8" to Github Actions
This commit adds a new GitHub Actions workflow to build the Vespa
systemtest base Docker image for CentOS 8. The workflow is triggered
on pushes to the main branch and pull requests targeting the main
branch, and it builds and pushes the Docker image to Docker Hub.
 image.
feat(ci): Schedule container image build in weekly cron