Add system test of JSONL output from Document V1
Also adds a JSONL convenience method to the
`lib/document_api_v1` module, which accepts a callback
handler for operations received as part of the JSONL
response. HTTP response reading is currently not done
in a streaming manner, but the API is designed to make
this transparently possible in the future, should it be
needed.